semntn/requirements.txt
click==8.1.7
contourpy==1.3.0
cycler==0.12.1
fonttools==4.55.3
joblib==1.4.2
kiwisolver==1.4.5
markdown-it-py==3.0.0
matplotlib==3.9.2
mdurl==0.1.2
numpy==1.26.4
packaging==24.1
pandas==2.2.2
pillow==10.4.0
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2024.2
PyYAML==6.0.2
rich==13.9.4
scikit-learn==1.5.2
scipy==1.14.1
shellingham==1.5.4
six==1.16.0
threadpoolctl==3.5.0
tqdm==4.66.5
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
pytest
